Fixed printf incorrect cast bug.

+++ src/arc/tty.c
@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2011 Graham Edgecombe <graham@grahamedgecombe.com>
+ *
+ * Permission to use, copy, modify, and/or distribute this software for any
+ * purpose with or without fee is hereby granted, provided that the above
+ * copyright notice and this permission notice appear in all copies.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+ * WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+ * MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+ * ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+ * WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+ * ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+ * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+ */
+
 #include <arc/tty.h>
 #include <arc/cpu/port.h>
 #include <arc/lock/spinlock.h>
@@ -378,7 +394,7 @@ void tty_vprintf(const char *fmt, va_list args)
             if (sign)
               value = (signed long int) value;
             else
-              value = (unsigned short int) value;
+              value = (unsigned long int) value;
           }
 
           /* convert the value to a string */
